add memcache implementation of httpcache.Cache
This provides two separate implementations, one that uses App Engine's
memcache service, and one that uses the gomemcache library to connect to
a specified memcached server.  The correct implementation is built based
on the presence of the "appengine" build constraint.

Tests for the non-App Engine implementation expect to connect to a
memcached server on localhost:11211 (the default memcached port).  If a
local server is not found, the tests are skipped.  Eventually this could
be replaced with a very simple TCP listener that implements enough of
the memcache protocol to run the tests.