Support multiple Vary response headers
If a server responds with multiple Vary headers, their comma-separated
field values should be joined and the joined list should be used to
determine whether a cached response may satisfy a request. (See
http://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.2.)

This commit adds headerAllCommaSepValues, which joins all
comma-separated values of a header, and uses it to compute the list of
Vary'd headers.