Change tests to use standard library, more idiomatic style.
The motivation for using idiomatic style and standard library for tests
is that it's more universal, and more familiar to a larger set of people.
It makes it easier to read, modify tests for experienced Go programmers.
It does not require thorough knowledge and familiarity with custom APIs.

Resolves #41.

This commit is largely based on previous work by @uovobw in
https://github.com/sourcegraph/httpcache/pull/3, modified by me to
improve error messages, style, and address preliminary review comments.