README: More consistent formatting with other Go packages. (#68)
Remove the redundant "A Transport for Go's http.Client that will
cache responses according to the HTTP RFC" sentence. The next
paragraph says something similar. This sentence can already be
found in the repository description (one liner).

It's very common to find a godoc.org link in the form of a badge in
most Go package repository READMEs. Therefore, it's more recognizable
and easier to click the link. I think we should use the badge to be
friendlier to users, and remove the superseded plain text link.

Move License note into its own section at the bottom. This is a more
common location. GitHub already detects and displays the license
prominently in top right corner, and there is a LICENSE file in the repo,
so it makes sense to place Cache Backends section higher up.