Reduce bufio.Reader buffering in newGatewayTimeoutResponse
bufio.NewReader uses a default buffer size of 4KiB
which is allocated up front. The entire response is
only 32-bytes in size, and thus the allocation is
entirely wasteful. By calling bufio.NewReaderSize
with a requested size of 0, we instead get the
minimum buffering size of 16-bytes. This is far less
wasteful and reduces memory allocations.