Cap bufio.Reader buffer size in CachedResponse
bufio.NewReader uses a default buffer size of 4KiB which is allocated
up front. As the size of the response is known ahead of time, limiting
the size of the buffer will reduce wasteful memory allocations.

It would be more efficient to eliminate buffering entirely, as the data
is already in memory, but there is no clear way to achieve this and
it's not clear that a lower buffer size would succeed in all cases.