httpcache: Fail fast in RoundTrip
The previous implementation was hard to audit for error checking.  For
example:

  if cacheable {
    cachedResp, err = CachedResponse(t.Cache, req)
  } else {
    ...
  }

  ...

  if cacheable && cachedResp != nil && err == nil {
    ...
  } else {
    reqCacheControl := parseCacheControl(req.Header)
    if _, ok := reqCacheControl["only-if-cached"]; ok {
      resp = newGatewayTimeoutResponse(req)
    } else {
      ...
    }
  }

  if cacheable && canStore(parseCacheControl(req.Header), parseCacheControl(resp.Header)) {
    ...
  } else {
    t.Cache.Delete(cacheKey)
  }
  return resp, nil

is a much harder path to trace than this commit's:

  cachedResp, err := CachedResponse(t.Cache, req)
  if err != nil {
    return nil, err
  }

The old approach allowed for a nil-error 504 despite the failed
CachedResponse.  That might be useful in some cases, but it masked a
broken cache backend.  With this commit, we surface that cache
breakage so users can fix it.