Streaming caches
For backends that support it, avoid holding the whole response body in
memory.  This is much more efficient when shipping around large files
on hosts without lots of memory.

For backends that don't support streaming (currently all except
diskcache), this is not much of a change.  The new
BufferedStreamingCache wrapper just collects the buffering logic in a
different place.

The streaming body is cached when the pipe is closed with a successful
"consumer read the whole body" EOF.  But because the handler is
running in a goroutine, we need a channel to wait out the handler exit
before continuing in the caller.  I initially only read from the
handlerErrors channel in streamReader.Close and left streamReader.Read
errors asynchronous.  But a number of the unit tests are structured:

  {
    // first request
    ...
    defer resp.Body.Close()
    ...
  }
  {
    // second request
  }

But defered functions do not fire on those scope exits:

  $ cat test.go
  package main

  import "fmt"

  func main() {
    {
      defer fmt.Println("a")
    }
    fmt.Println("b")
  }
  $ go run test.go
  b
  a
  $ go version
  go version go1.10.3 linux/amd64

You could force the defers to fire with:

  func() {
    // first request
    ...
    defer resp.Body.Close()
    ...
  }()

but adding channel readers to both Read() and Close() ensures we sync
with the handler exit on either EOF or early-close.