.travis.yml: Only run 'gofmt' on 1.11.x
The indent heuristic changed with Go 1.11 [1], so there's no single
form that satisfies gofmt for both Go 1.10 and Go 1.11.  With this
commit, we now only run gofmt on 1.11.

Docs for the 'include' syntax are in [2].

Also replace the double-diff with a 'gofmt -w ...' to write the
changes and a 'git diff ...' call to display the changes and exit
non-zero if there were any.

[1]: https://golang.org/doc/go1.11#gofmt
[2]: https://docs.travis-ci.com/user/customizing-the-build/#explicitly-including-jobs