.travis.yml: bump to Go 1.11, use 'gofmt -s' (#92)
* .travis.yml: Bump to modern Go (1.11)

Go 1.11 came out in August [1], which means Go 1.9 and earlier are no
longer supported [2].  Just test on supported versions.

[1]: https://golang.org/doc/devel/release.html#go1.11
[2]: https://golang.org/doc/devel/release.html#policy

* Revert "Remove simplify-code option, for now"

This reverts commit be6978e2943923f8ece88a9a23d6d7a343edeeef (Remove
simplify-code option, for now, 2016-03-12, #40).  I expect this was
due to a simplification change between Go versions (the commit message
doesn't say), but none of the versions which we were using back then
are supported any more.

* httpcache: Update gofmt

Generated with:

  $ gofmt -s -w *.go */*.go

using:

  $ go version
  go version go1.11.2 linux/amd64

* .travis.yml: Only run 'gofmt' on 1.11.x

The indent heuristic changed with Go 1.11 [1], so there's no single
form that satisfies gofmt for both Go 1.10 and Go 1.11.  With this
commit, we now only run gofmt on 1.11.

Docs for the 'include' syntax are in [2].

Also replace the double-diff with a 'gofmt -w ...' to write the
changes and a 'git diff ...' call to display the changes and exit
non-zero if there were any.

[1]: https://golang.org/doc/go1.11#gofmt
[2]: https://docs.travis-ci.com/user/customizing-the-build/#explicitly-including-jobs